Consolidate vitest environment tests into vitest-config.test.ts
Merge tests from vitest-env-setup.test.ts into vitest-config.test.ts to
eliminate duplicate vitest config loading and improve test organization.
Replace raw fs/path usage with existing loadProjectFile helper for
consistency. Mark task 3 subtasks as complete in spec.
