+++ .kiro/specs/vitest-deck-tests/prompt.md
@@ -45,6 +45,7 @@ think.
 
 takt --task "/kiro:spec-impl vitest-deck-tests 1"
 takt --task "/kiro:spec-impl vitest-deck-tests 2"
+takt --task "/kiro:spec-impl vitest-deck-tests 3"
