+++ .kiro/specs/vitest-deck-tests/tasks.md
@@ -57,12 +57,12 @@
   - `shuffleDeck()` がシャッフル後も52枚を維持することを検証する
   - _Requirements: 3.1_
 
-- [ ] 3. 全テストの動作確認を行う
-- [ ] 3.1 単体テストを実行して全テストがパスすることを確認する
+- [x] 3. 全テストの動作確認を行う
+- [x] 3.1 単体テストを実行して全テストがパスすることを確認する
   - `npm run test` を実行し、新規テストと既存テストが全てパスすることを確認する
   - _Requirements: 1.1, 1.2, 1.3, 1.4, 1.6, 2.1, 2.2, 2.3, 2.4, 3.1_
 
-- [ ] 3.2 E2Eテストを実行してUIに影響がないことを確認する
+- [x] 3.2 E2Eテストを実行してUIに影響がないことを確認する
   - `npm run test:e2e` を実行し、スクリーンショット差分が0であることを確認する
   - UI変更がないため、既存のベースラインとの差分がないことを検証する
   - _Requirements: 1.1, 1.2, 1.3, 1.4, 1.5, 1.6, 2.1, 2.2, 2.3, 2.4, 3.1_
